[Http] Refactor Stream.
